Desktop/RunTimers_final_updated/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
openai>=1.30.0
requests>=2.31.0
python-multipart>=0.0.9
pyyaml>=6.0
pdfplumber>=0.10.0
